feat: add search utility functions and types for improved search functionality
